refactor: validator
