+++ src/routes/api/Session.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 
 import SessionController from '../../controllers/api/Session';
 import { asyncMiddleware } from '../../plugins/utility';
-import SessionValidator from '../../validator/util';
+import SessionValidator from '../../validator/api/index';
 
 const router = express.Router();
 const { create, findById, update, destroy, find } = SessionController;
