+++ src/routes/api/Transaction.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 
 import controller from '../../controllers/api/Transaction';
 import { asyncMiddleware } from '../../plugins/utility';
-import Validator from '../../validator/util';
+import Validator from '../../validator/api/index';
 
 const router = express.Router();
 const { create, update, findByid, destroy, find } = controller;
