+++ src/routes/api/User.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 
 import controller from '../../controllers/api/User';
 import { asyncMiddleware } from '../../plugins/utility';
-import Validator from '../../validator/util';
+import Validator from '../../validator/api/index';
 
 const router = express.Router();
 const { create, update, destroy, findById, find } = controller;
