+++ src/routes/craw/Transaction.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 
 import controller from '../../controllers/job/Transactions';
 import { asyncMiddleware } from '../../plugins/utility';
-import CrawlValidator from '../../validator/util';
+import CrawlValidator from '../../validator/function/index';
 
 const router = express.Router();
