+++ src/validator/api/Stock.ts
